+++ src/main/java/roomescape/exception/handler/GlobalExceptionHandler.java
@@ -1,6 +1,5 @@
-package roomescape.exception;
+package roomescape.exception.handler;
 
-import jakarta.servlet.annotation.HttpConstraint;
 import java.util.NoSuchElementException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -9,6 +8,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 import roomescape.dto.response.ErrorResponse;
+import roomescape.exception.custom.PastReservationTimeException;
 
 @RestControllerAdvice
 @Slf4j
