+++ src/main/java/roomescape/exception/handler/ValidationExceptionHandler.java
@@ -1,0 +1,4 @@
+package roomescape.exception.handler;
+
+public class ValidationExceptionHandler {
+}
